integrated docker

+++ server/CodeExecuter/codeExecuter.js
@@ -5,7 +5,6 @@ const { exec, spawn } = require("child_process");
 
 const codeDirectory = path.join(__dirname, "codeFiles");
 const outputDirectory = path.join(__dirname, "outputFiles");
-const testCasesDirectory = path.join(__dirname, "testcases");
 
 // for the first time create 'codeFiles' directory
 if (!fs.existsSync(codeDirectory)) {
@@ -26,11 +25,17 @@ const createFile = (fileExtension, content) => {
 }
 
 const deleteFile = filepath => {
-    const id = path.basename(filepath).split(".")[0];
-    const outPath = path.join(outputDirectory, `${id}.out`);
-
+    if (!fs.existsSync(filepath)) return;
     fs.unlinkSync(filepath);
+    console.log('unlinked : ', filepath);
+}
+
+const deleteOutFile = (filepath, extension) => {
+    const id = path.basename(filepath).split(".")[0];
+    const outPath = path.join(outputDirectory, `${id}.${extension}`);
+    if (!fs.existsSync(outPath)) return;
     fs.unlinkSync(outPath);
+    console.log('unlinked : ', outPath);
 }
 
 const execCppCode = (filePath, testcase) => {
@@ -42,7 +47,7 @@ const execCppCode = (filePath, testcase) => {
         exec(
             `g++ "${filePath}" -o "${outPath}"`,
             (error, stdout, stderr) => {
-                error && reject({ msg: 'on error', error, stderr });
+                error && reject({ msg: 'on error', error: `${error.name} => ${error.message}`, stderr });
                 stderr && reject({ msg: 'on stderr', stderr });
 
                 let passed = 0;
@@ -52,7 +57,7 @@ const execCppCode = (filePath, testcase) => {
                     cmd.stdin.write(testInput);
                     cmd.stdin.end();
                     cmd.on('error', error => {
-                        reject({ msg: 'on error', error });
+                        reject({ msg: 'on error', error: `${error.name} => ${error.message}` });
                     });
                     cmd.stderr.on('data', data => {
                         reject({ msg: 'on stderr', stderr: `${data}` });
@@ -61,14 +66,13 @@ const execCppCode = (filePath, testcase) => {
                         const exOut = `${data}`.trim();
                         if (exOut === output[index]) {
                             passed += 1;
-                            console.log(`testcase ${index} passed !`);
+                            // if socket connection established then send to client the index of passed test case
                             (passed === input.length) && resolve({ msg: 'All Test Cases Passed' });
                         } else {
                             reject({ msg: 'on wrong answer', stderr: `testcase ${index} failed => testcase : (${testInput}) => expected : ${output[index]} => your output : ${exOut}` });
                         }
                     });
                     cmd.on('close', code => {
-                        console.log(`child process exited with code ${code}`);
                     });
                 });
 
@@ -77,8 +81,45 @@ const execCppCode = (filePath, testcase) => {
     });
 }
 
+
+const execPyCode = (filePath, testcase) => {
+    let { input, output } = require(`./testcases/${testcase}`)
+
+    return new Promise((resolve, reject) => {
+        let passed = 0;
+        input.forEach((testInput, index) => {
+            testInput = testInput.replaceAll(' ', '\n');
+
+            const cmd = spawn(`python "${filePath}"`, [], { shell: true })
+            cmd.stdin.write(testInput);
+            cmd.stdin.end();
+            cmd.on('error', error => {
+                reject({ msg: 'on error', error: `${error.name} => ${error.message}` });
+            });
+            cmd.stderr.on('data', data => {
+                reject({ msg: 'on stderr', stderr: `${data}` });
+            });
+            cmd.stdout.on('data', data => {
+                const exOut = `${data}`.trim();
+                if (exOut === output[index]) {
+                    passed += 1;
+                    // if socket connection established then send to client the index of passed test case
+                    (passed === input.length) && resolve({ msg: 'All Test Cases Passed' });
+                } else {
+                    reject({ msg: 'on wrong answer', stderr: `testcase ${index} failed => testcase : (${testInput}) => expected : ${output[index]} => your output : ${exOut}` });
+                }
+            });
+            cmd.on('close', code => {
+                console.log(`child closed with code ${code}`);
+            });
+        });
+    });
+}
+
 module.exports = {
     createFile,
+    execCppCode,
+    execPyCode,
     deleteFile,
-    execCppCode
+    deleteOutFile
 };
